+++ aws-distro-opentelemetry-node-autoinstrumentation/src/aws-span-exporter/README.md
