+++ aws-distro-opentelemetry-node-autoinstrumentation/src/aws-span-exporter/otlp-aws-span-exporter.ts
